서버디비에서 내부디비로 이전작업. Folder 보여주기 부분 완료. simpletodo 생성, 보여주기 부분 완료.

+++ app/src/main/java/com/team/codealmanac/w2do/models/TodoFolder.java
@@ -6,13 +6,13 @@ package com.team.codealmanac.w2do.models;
 
 public class TodoFolder {
     public String name;
-    public int sequence;
-    public int todo_count;
+    public long sequence;
+    public long todo_count;
 
     public TodoFolder() {
     }
 
-    public TodoFolder(String name, int sequence, int todo_count) {
+    public TodoFolder(String name, long sequence, long todo_count) {
         this.name = name;
         this.sequence = sequence;
         this.todo_count = todo_count;
